Move StrptimeParser and Java related parsers to a dedicated file
This way it is easier to keep it manageable and clearly separate QW API DatParser logic and ES API related parser.

Tests moved together with the code base.
