+++ quickwit/quickwit-datetime/src/lib.rs
@@ -19,9 +19,11 @@
 
 mod date_time_format;
 mod date_time_parsing;
+pub mod java_date_time_format;
 
-pub use date_time_format::{DateTimeInputFormat, DateTimeOutputFormat, StrptimeParser};
+pub use date_time_format::{DateTimeInputFormat, DateTimeOutputFormat};
 pub use date_time_parsing::{
     parse_date_time_str, parse_timestamp, parse_timestamp_float, parse_timestamp_int,
 };
+pub use java_date_time_format::StrptimeParser;
 pub use tantivy::DateTime as TantivyDateTime;
